feat: create new project config for invalid licenses
